fix(cloud): isolate funding tests and use database time
